requirements.txt
feedparser==6.0.11
python-dateutil==2.9.0.post0
Unidecode==1.3.8